got the netowrk-switcher working
