+++ utils/switcher/index.html
@@ -1,26 +1,11 @@
-<!DOCTYPE html>
-<!--[if lt IE 7]>      <html class="no-js lt-ie9 lt-ie8 lt-ie7"> <![endif]-->
-<!--[if IE 7]>         <html class="no-js lt-ie9 lt-ie8"> <![endif]-->
-<!--[if IE 8]>         <html class="no-js lt-ie9"> <![endif]-->
-<!--[if gt IE 8]><!--> <html class="no-js"> <!--<![endif]-->
+<html>
 <head>
   <meta charset="utf-8">
-  <meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1">
   <title>Playlist Switching Simulator</title>
-  <meta name="description" content="">
-  <meta name="viewport" content="width=device-width">
-
-  <link rel="stylesheet" href="css/normalize.min.css">
-  <link rel="stylesheet" href="css/main.css">
-  <link rel="stylesheet" href="../../node_modules/video.js/dist/video-js.css">
-
-  <script src="js/vendor/modernizr-2.6.2.min.js"></script>
+  <link rel="stylesheet" href="switcher.css">
+  <link rel="stylesheet" href="/node_modules/video.js/dist/video-js.css">
 </head>
 <body>
-  <!--[if lt IE 7]>
-      <p class="chromeframe">You are using an <strong>outdated</strong> browser. Please <a href="http://browsehappy.com/">upgrade your browser</a> or <a href="http://www.google.com/chromeframe/?redirect=true">activate Google Chrome Frame</a> to improve your experience.</p>
-      <![endif]-->
-
   <div class="header-container">
     <header class="wrapper clearfix">
       <h1 class="title">Playlist Switching Simulator</h1>
@@ -29,7 +14,6 @@
 
   <div class="main-container">
     <div class="main wrapper clearfix">
-
       <article>
         <header>
           <p>
@@ -37,11 +21,6 @@
             playlist switcher under different network conditions. It
             can helpful to understand how tweaks to the switching
             logic will affect playback.
-          </p>
-          <p>
-            Flash security restrictions prevent this page from running
-            over the file protocol. Run <code>grunt connect</code> and
-            then <a href="http://localhost:9999/test/switcher">reload.</a>
           </p>
         </header>
         <section>
@@ -58,7 +37,7 @@
                   After <input name=time0 class=time type=number value=0 min=0> seconds,
                 </label>
                 <label>
-                  the link capacity is <input name=bandwidth0 class=bandwidth type=number value=921600> bits per second
+                  the link capacity is <input name=bandwidth0 class=bandwidth type=number value=250000> bits per second
                 </label>
               </li>
               <li>
@@ -66,7 +45,7 @@
                   After <input name=time1 class=time type=number value=150 min=0> seconds,
                 </label>
                 <label>
-                  the link capacity is <input name=bandwidth1 class=bandwidth type=number value=450560> bits per second
+                  the link capacity is <input name=bandwidth1 class=bandwidth type=number value=500000> bits per second
                 </label>
               </li>
               <li>
@@ -74,7 +53,7 @@
                   After <input name=time2 class=time type=number value=600 min=0> seconds,
                 </label>
                 <label>
-                  the link capacity is <input name=bandwidth2 class=bandwidth type=number value=1843200> bits per second
+                  the link capacity is <input name=bandwidth2 class=bandwidth type=number value=250000> bits per second
                 </label>
               </li>
             </ol>
@@ -94,44 +73,28 @@
             </p>
             The video is available at
             <ul>
-              <li><input class=bitrate type=number min=1 value=312000> bits per second</li>
-              <li><input class=bitrate type=number min=1 value=524000> bits per second</li>
-              <li><input class=bitrate type=number min=1 value=1296000> bits per second</li>
-              <li><input class=bitrate type=number min=1 value=2125000> bits per second</li>
-              <li><input class=bitrate type=number min=1 value=3125000> bits per second</li>
+              <li><input class=bitrate type=number min=1 value=52400> bits per second</li>
+              <li><input class=bitrate type=number min=1 value=129600> bits per second</li>
+              <li><input class=bitrate type=number min=1 value=212500> bits per second</li>
+              <li><input class=bitrate type=number min=1 value=312500> bits per second</li>
             </ul>
           </form>
         </section>
       </article>
 
-    </div> <!-- #main -->
-  </div> <!-- #main-container -->
+    </div>
+  </div>
 
   <div class="footer-container">
     <footer class="wrapper">
       <h3>videojs-contrib-hls</h3>
     </footer>
   </div>
-  <div id=fixture></div>
-  <script src="../../node_modules/sinon/lib/sinon.js"></script>
-  <script src="../../node_modules/sinon/lib/sinon/util/event.js"></script>
-  <script src="../../node_modules/sinon/lib/sinon/util/fake_xml_http_request.js"></script>
-  <script src="../../node_modules/sinon/lib/sinon/util/xhr_ie.js"></script>
-  <script src="../../node_modules/sinon/lib/sinon/util/fake_timers.js"></script>
-  <script src="js/vendor/d3.min.js"></script>
-
+  <div id="qunit-fixture"></div>
+  <script src="/node_modules/sinon/pkg/sinon.js"></script>
   <script src="/node_modules/video.js/dist/video.js"></script>
-  <script src="/node_modules/videojs-contrib-media-sources/dist/videojs-media-sources.js"></script>
-  <script src="/node_modules/pkcs7/dist/pkcs7.unpad.js"></script>
-  <script src="/src/videojs-hls.js"></script>
-  <script src="/src/xhr.js"></script>
-  <script src="/src/stream.js"></script>
-  <script src="/src/m3u8/m3u8-parser.js"></script>
-  <script src="/src/playlist.js"></script>
-  <script src="/src/playlist-loader.js"></script>
-  <script src="/src/decrypter.js"></script>
-  <script src="/src/bin-utils.js"></script>
-
-  <script src="js/switcher.js"></script>
+  <script src="/dist/videojs-contrib-hls.js"></script>
+  <script src="/node_modules/d3/d3.min.js"></script>
+  <script src="/dist/switcher.js"></script>
 </body>
 </html>
