--- utils/switcher/js/main.js
@@ -1,1 +1,0 @@
-
